requirements.txt
absl-py==0.7.1
astor==0.8.0
cycler==0.10.0
emcee==2.2.1
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
kiwisolver==1.1.0
Markdown==3.0.1
matplotlib==2.2.4
numpy==1.14.5
protobuf==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
scipy==1.2.2
six==1.12.0
tensorboard==1.10.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
Werkzeug==0.15.4
